# api/openapi.yaml
openapi: 3.1.0
info:
  title: AgentOS API
  version: 0.1.0
  summary: Multi-tenant AI-agent infrastructure platform API
  description: |
    AgentOS is a multi-tenant AI-agent infrastructure platform. This contract
    covers the versioned REST API (/api/v1), the worker pull endpoint, the
    Server-Sent Events (SSE) execution-event stream, and platform health /
    metrics endpoints.

    ## Authentication
    Two schemes are accepted on every protected endpoint (see
    `securitySchemes`):
    1. `bearerToken` — `Authorization: Bearer <token>` where `<token>` is the
       HMAC-SHA256 JWT-style token returned by `POST /auth/login`
       (registration creates the first OWNER user; log in to get a token).
    2. `apiKey` — `X-API-Key: ak_...` header. API keys are created server-side
       (dev key created at startup) and authenticate as the key's OWNER role.
    Browser `EventSource` clients that cannot set headers may pass
    `?api_key=<key>` as a query parameter (legacy compatibility).

    ## RBAC
    Permissions: `agents.read`, `agents.write`, `runs.read`, `runs.execute`,
    `users.manage`, `organization.manage`. Roles: `OWNER` (all), `ADMIN`
    (all but organization.manage), `MEMBER` (agents.read, runs.read,
    runs.execute), `VIEWER` (agents.read, runs.read). Each operation declares
    its required permission via `x-required-permission`.

    ## Multi-tenancy
    Requests are scoped to an organization derived from the auth token. For
...

    ## Error model
...

...
servers:
...
security:
- bearerToken: []
- apiKey: []
paths:
...
  /policies:
    get:
      operationId: listPolicies
      summary: List policies of the authenticated organization
      description: |
        Returns the governance policies of the caller's organization ordered
        by priority (desc) then creation time (asc) — the same order the
        evaluation engine applies. Unpaginated (see base spec limitation).
      tags:
      - Governance
      security:
      - bearerToken: []
      - apiKey: []
      x-required-permission: policies.read
      responses:
        '200':
          description: Policy list (possibly empty)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/Internal'
  /policies/create:
    post:
      operationId: createPolicy
      summary: Create a policy
      description: |
        Validates the policy record (name required; effect allow|deny;
        resource_type tool|agent|workflow|deployment|*; no empty entries in
        actions / tool_allowlist / environments; max_cost_cents >= 0) and
        persists it for the caller's organization. id and timestamps are
        assigned server-side. Returns 422 with an invalid_policy error when
        validation fails.
      tags:
      - Governance
      security:
      - bearerToken: []
      - apiKey: []
      x-required-permission: policies.write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Policy'
      responses:
        '201':
          description: Created policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '400':
          description: Malformed JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Policy validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /policies/evaluate:
    post:
      operationId: evaluatePolicies
      summary: Evaluate the caller's policies against one action
      description: |
        Runs the governance engine over the enabled policies of the caller's
        organization: highest-priority match wins, deny beats allow on a
        priority tie, and the default decision is allow when nothing matches.
        Condition semantics (conditions are applicability predicates):
          * tool_allowlist (non-empty) — the request tool must be listed;
          * environments (non-empty) — the request environment must be listed;
          * max_cost_cents (set >= 0) — the request is over budget, i.e.
            estimated_cost_cents > max_cost_cents (budget-guard);
          * require_approval — annotated in the decision reason.
        The tenant scope always comes from the authenticated principal;
        resource.tenant_id is informational. Evaluate is a read-only operation
        and requires policies.read.
      tags:
      - Governance
      security:
      - bearerToken: []
      - apiKey: []
      x-required-permission: policies.read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyEvaluateRequest'
      responses:
        '200':
          description: Decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyEvaluateResponse'
        '400':
          description: Malformed JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Missing action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /policies/{id}:
    get:
      operationId: getPolicy
      summary: Get one policy
      tags:
      - Governance
      security:
      - bearerToken: []
      - apiKey: []
      x-required-permission: policies.read
      parameters:
      - $ref: '#/components/parameters/PolicyID'
      responses:
        '200':
          description: The policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updatePolicy
      summary: Full update of one policy
      description: |
        Replaces the mutable fields (name, effect, resource_type, actions,
        conditions, priority, enabled). id, organization_id and created_at
        are preserved; updated_at is bumped. Validates like create.
      tags:
      - Governance
      security:
      - bearerToken: []
      - apiKey: []
      x-required-permission: policies.write
      parameters:
      - $ref: '#/components/parameters/PolicyID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Policy'
      responses:
        '200':
          description: Updated policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '400':
          description: Malformed JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Policy validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deletePolicy
      summary: Delete one policy
      tags:
      - Governance
      security:
      - bearerToken: []
      - apiKey: []
      x-required-permission: policies.write
      parameters:
      - $ref: '#/components/parameters/PolicyID'
      responses:
        '200':
          description: Deletion confirmation
          content:
            application/json:
              schema:
                type: object
                required:
                - deleted
                properties:
                  deleted:
                    type: boolean
                    enum:
                    - true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: HS256-HMAC
      description: |
        HMAC-SHA256 JWT-style token from `POST /auth/login` (24h expiry).
        Claims: user_id, organization_id, email, role, exp.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Pre-provisioned API key (`ak_...`). Authenticates as the key's
        organization with the OWNER role. Browser EventSource clients may
        alternatively pass `?api_key=<key>` (legacy query-param auth).
  responses:
    Unauthorized:
      description: Missing/invalid credentials (plain text body today; JSON model is the target).
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/LegacyTextError'
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but lacking the required permission or organization access.
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/LegacyTextError'
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found.
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/LegacyTextError'
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    LegacyTextError:
      type: string
      description: Legacy plain-text error body emitted via http.Error during migration.
      example: agent not found
    Error:
      type: object
      description: Structured error model (internal/httpx.ErrorModel).
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable stable error code (e.g. unauthorized, forbidden, not_found,
                bad_request, internal_error).
              examples:
              - not_found
            message:
              type: string
              description: Human-readable message.
              examples:
              - agent not found
            request_id:
              type: string
              description: Correlation ID echoed from the X-Request-ID header (or generated per request).
              examples:
              - 1b8f0c47e51d2a3c4d5e6f708192a3b4
          required:
          - code
          - message
          - request_id
      required:
      - error
...
    PolicyConditions:
      type: object
      description: |
        Applicability predicates for a policy. An absent condition does not
        restrict the policy. See POST /policies/evaluate for the semantics.
      properties:
        tool_allowlist:
          type: array
          description: Restricts the policy to the listed tool ids/names.
          items:
            type: string
        max_cost_cents:
          oneOf:
          - type: integer
            minimum: 0
          - type: 'null'
          description: |
            When set, the policy applies to requests whose estimated cost
            exceeds this cap (budget-guard condition).
        environments:
          type: array
          description: Restricts the policy to the listed environments.
          items:
            type: string
            enum:
            - development
            - staging
            - production
        require_approval:
          type: boolean
          description: |
            Not a match predicate; when the winning policy carries it the
            decision reason notes that approval is required.
    Policy:
      type: object
      required:
      - name
      - effect
      - resource_type
      - actions
      - conditions
      - priority
      - enabled
      - created_at
      - updated_at
      properties:
        id:
          type: string
          readOnly: true
        name:
          type: string
        effect:
          type: string
          enum:
          - allow
          - deny
        resource_type:
          type: string
          enum:
          - tool
          - agent
          - workflow
          - deployment
          - '*'
        actions:
          type: array
          description: |
            Actions this policy governs, e.g. ["runs.execute","tools.call"].
            An empty list matches every action.
          items:
            type: string
        conditions:
          $ref: '#/components/schemas/PolicyConditions'
        priority:
          type: integer
          description: Higher wins; ties resolve deny-over-allow.
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
          description: RFC3339 UTC.
        updated_at:
          type: string
          format: date-time
          description: RFC3339 UTC.
    PolicyResponse:
      type: object
      required:
      - policy
      properties:
        policy:
          $ref: '#/components/schemas/Policy'
    PolicyListResponse:
      type: object
      required:
      - policies
      properties:
        policies:
          type: array
          items:
            $ref: '#/components/schemas/Policy'
    PolicyEvaluateRequest:
      type: object
      required:
      - action
      - resource
      - context
      properties:
        subject:
          type: object
          description: Informational caller identity.
          properties:
            user_id:
              type: string
            role:
              type: string
            api_key_id:
              type: string
        action:
          type: string
          description: The governed action, e.g. tools.call or runs.execute.
        resource:
          type: object
          properties:
            type:
              type: string
              enum:
              - tool
              - agent
              - workflow
              - deployment
              - '*'
            id:
              type: string
            tenant_id:
              type: string
              description: Informational; the actual scope comes from auth.
        context:
          type: object
          properties:
            estimated_cost_cents:
              type: integer
              description: Estimated cost used by the max_cost_cents condition.
            environment:
              type: string
              description: Target environment used by the environments condition.
            tool:
              type: string
              description: Overrides resource.id for tool_allowlist matching.
    PolicyEvaluateResponse:
      type: object
      required:
      - decision
      - matched_policy_id
      - reason
      properties:
        decision:
          type: string
          enum:
          - allow
          - deny
        matched_policy_id:
          type: string
          description: Empty when no policy matched (default allow).
        reason:
          type: string
...